Moved SponsorLinksHelper logic to ToString overload on model
