+++ src/partials/Program.Sponsors.cs
@@ -6,34 +6,6 @@ namespace Pannella;
 
 internal partial class Program
 {
-    private static string SponsorLinksHelper(Sponsor sponsor)
-    {
-        var links = new StringBuilder();
-
-        var properties = typeof(Sponsor).GetProperties();
-
-        foreach (var prop in properties)
-        {
-            object value = prop.GetValue(sponsor, null);
-            if (value != null && value.GetType() == typeof(List<string>))
-            {
-                //String.Join didn't want to work unless I explicitely cast this
-                var stringArray = value as List<string>;
-                links.AppendLine();
-                links.AppendLine();
-                links.Append(string.Join(Environment.NewLine, stringArray));
-            }
-            else if (value != null && value.GetType() == typeof(string))
-            {
-                links.AppendLine();
-                links.AppendLine();
-                links.Append(value);
-            }
-        }
-
-        return links.ToString();
-    }
-
     private static string GetRandomSponsorLinks()
     {
         var output = new StringBuilder();
@@ -46,12 +18,11 @@ internal partial class Program
 
             if (randomItem.sponsor != null)
             {
-                var links = SponsorLinksHelper(randomItem.sponsor);
                 var author = randomItem.GetConfig().metadata.author;
 
                 output.AppendLine();
                 output.AppendLine($"Please consider supporting {author} for their work on the {randomItem} core:");
-                output.Append(links.Trim());
+                output.Append(randomItem.sponsor);
             }
         }
 
@@ -85,11 +56,9 @@ internal partial class Program
         {
             if (core.sponsor != null)
             {
-                var links = SponsorLinksHelper(core.sponsor);
-
                 Console.WriteLine();
                 Console.WriteLine($"{core.identifier}:");
-                Console.WriteLine(links.Trim());
+                Console.WriteLine(core.sponsor);
             }
         }
     }
